remove foundry chain
enabled using local AA infra by setting env variables
