+++ apps/laboratory/.env.example
@@ -13,3 +13,5 @@ NEXT_PUBLIC_DEFAULT_LOG_LEVEL=""
 NEXT_PUBLIC_PIMLICO_KEY=""
 # local env setup guide  https://docs.pimlico.io/permissionless/how-to/local-testing"
 NEXT_PUBLIC_LOCAL_BUNDLER_URL=
+NEXT_PUBLIC_LOCAL_PAYMASTER_URL=
+NEXT_PUBLIC_LOCAL_CLIENT_URL=
